requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
abseil-cpp=20211102.0=h27087fc_1
aiohttp=3.8.1=py310h5764c6d_1
aiosignal=1.3.1=pyhd8ed1ab_0
appdirs=1.4.4=pyhd3eb1b0_0
arrow-cpp=11.0.0=h374c478_2
async-timeout=4.0.3=pyhd8ed1ab_0
attrs=23.1.0=pyh71513ae_1
aws-c-common=0.6.8=h5eee18b_1
aws-c-event-stream=0.1.6=h6a678d5_6
aws-checksums=0.1.11=h5eee18b_2
aws-sdk-cpp=1.8.185=h721c034_1
blas=1.0=mkl
boost-cpp=1.78.0=he72f1d9_0
bottleneck=1.3.5=py310ha9d4c09_0
brotli-python=1.0.9=py310h6a678d5_7
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=h5eee18b_0
ca-certificates=2023.7.22=hbcca054_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=unix_pyh707e725_0
colorama=0.4.6=pyhd8ed1ab_0
cryptography=41.0.3=py310hdda0065_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cudatoolkit=11.4.1=h8ab8bb3_9
dataclasses=0.8=pyhc8e2a94_3
datasets=2.14.6=pyhd8ed1ab_0
dill=0.3.7=pyhd8ed1ab_0
faiss-gpu=1.7.4=py3.10_hc0239a3_0_cuda11.4
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py310h06a4308_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=py310h5eee18b_0
fsspec=2023.10.0=pyhca7485f_0
gflags=2.2.2=he1b5a44_1004
giflib=5.2.1=h5eee18b_3
glog=0.5.0=h48cff8f_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
grpc-cpp=1.48.2=he1ff14a_1
hnswlib=0.7.0=py310hcb52e73_2
huggingface_hub=0.18.0=pyhd8ed1ab_0
icu=70.1=h27087fc_0
idna=3.4=py310h06a4308_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
jinja2=3.1.2=py310h06a4308_0
joblib=1.3.2=pyhd8ed1ab_0
jpeg=9e=h5eee18b_1
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.8.0.34=0
libcurand=10.3.4.52=0
libcurl=8.4.0=h251f7ec_0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h516909a_1
libevent=2.1.12=hdbd6064_1
libfaiss=1.7.4=h13c3c6d_0_cuda11.4
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.2.0=h807b86a_2
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnghttp2=1.57.0=h2d74bed_0
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=13.2.0=h7e041cc_2
libtasn1=4.19.0=h5eee18b_0
libthrift=0.15.0=h1795dd8_2
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
llvm-openmp=14.0.6=h9e868ea_0
llvmlite=0.41.1=py310he1b5a44_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py310h7f8727e_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
multidict=6.0.2=py310h5eee18b_0
multiprocess=0.70.15=py310h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py310h06a4308_0
nltk=3.8.1=pyhd8ed1ab_0
numba=0.58.1=np1.21py3.10ha54aa7d_gd4460feb8_0
numexpr=2.8.4=py310h8879344_0
numpy=1.24.3=py310hd5efca6_0
numpy-base=1.24.3=py310h8e6c178_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.1.4=hd590300_0
orc=1.7.4=hb3bc3d3_1
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.1=py310h1128e8f_0
pillow=10.0.1=py310ha6cbd5a_0
pip=23.3=py310h06a4308_0
pooch=1.4.0=pyhd3eb1b0_0
pyarrow=11.0.0=py310h468efa6_1
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=py310h06a4308_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.13=h955ad1f_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd8ed1ab_0
python-xxhash=2.0.2=py310h5eee18b_1
python_abi=3.10=2_cp310
pytorch=2.0.1=py3.10_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pyhd8ed1ab_0
pyyaml=6.0=py310h5764c6d_4
re2=2022.04.01=h27087fc_0
readline=8.2=h5eee18b_0
regex=2023.10.3=py310h5eee18b_0
requests=2.31.0=py310h06a4308_0
sacremoses=0.0.53=pyhd8ed1ab_0
safetensors=0.4.0=py310ha89cbab_0
scikit-learn=1.3.0=py310h1128e8f_0
scipy=1.10.1=py310hd5efca6_0
sentence-transformers=2.2.2=pyhd8ed1ab_0
sentencepiece=0.1.99=py310hdb19cb5_0
setuptools=68.0.0=py310h06a4308_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.9=h295c915_0
sqlite=3.41.2=h5eee18b_0
sympy=1.11.1=py310h06a4308_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=py310h22610ee_0
torchaudio=2.0.2=py310_cu117
torchtriton=2.0.0=py310
torchvision=0.15.2=py310_cu117
tqdm=4.66.1=pyhd8ed1ab_0
transformers=4.31.0=pyhd8ed1ab_0
typing-extensions=4.7.1=py310h06a4308_0
typing_extensions=4.7.1=py310h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.18=py310h06a4308_0
utf8proc=2.6.1=h27cfd23_0
wheel=0.41.2=py310h06a4308_0
xxhash=0.8.0=h7f98852_3
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7f98852_2
yarl=1.7.2=py310h5764c6d_2
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
